backend/pyproject.toml
[project]
name = "obsidian-voice-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "whoosh>=2.7.4",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-frontmatter>=1.0.0",
    "pydantic-ai>=0.1",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "python-dotenv>=1.2.1",
    "livekit-agents[silero,turn-detector,openai]~=1.3",
    "livekit-api>=1.0.7,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
